Small polish update
Updated file sizes for all image,gifs and videos to make them smaller and less than 100mb total now is 40mb for everything. Updated refrence page so when you click on the link it opens a new page instead of using the same page so the music can keep on playuign in the background. and updated the refrence music from a wav file which is bigger to a mp3 which is smaller and more polished 320kb audio.